Update and rename app.py to mainpageapp.py
